+++ src/func/watch.sh
@@ -80,10 +80,8 @@ watch_Template()
 	watch_First
 	WATCH_LINES=$(tput lines)
 	STAT_AMOUNT=$(watch_Amount)
-	# create alternate screen buffer
-	printf "\e[?1049h"
-	# hide user input
-	stty -echo
+	# create alternate screen buffer + hide user input
+	[[ $FIRST_WATCH = true ]] && stty -echo && printf "\e[?1049h"
 	trap 'stty echo; printf "\e[?1049l\e[1;97m%s\e[1;95m%s\e[1;97m%s\n" "[Exiting: " "${SERVICE}" "]"; exit 0' EXIT
 
 	# need sudo for xmrig journals
@@ -98,7 +96,7 @@ watch_Template()
 				*"Active: failed"*) [[ $STAT_UPTIME = ... ]] && DOT_COLOR="\e[1;91mFAILED: ${NAME_PRETTY} $NAME_VER" || DOT_COLOR="\e[1;92mONLINE \e[1;93m(non-systemd): ${NAME_PRETTY} $NAME_VER";;
 				*) DOT_COLOR="\e[1;93m???: ${NAME_PRETTY}";;
 			esac
-			printf "\e[2J\e[H"
+			clear
 			echo -e "$STATS"
 			printf "\n\e[${WATCH_LINES};0H\e[1;97m[${DOT_COLOR}\e[1;97m] [\e[1;95m%s\e[1;97m%s\e[1;94m%s\e[1;97m%s\e[0;97m%s\e[1;97m%s\e[0;97m%s\e[1;97m%s\e[0m " \
 				"$STAT_DATE" "] [" "$STAT_UPTIME" "] [${WATCH_REFRESH_RATE} sec] [" "$STAT_AMOUNT" "]"
@@ -127,8 +125,8 @@ watch_Template()
 				*"Active: failed"*) [[ $STAT_UPTIME = ... ]] && DOT_COLOR="\e[1;91mFAILED: ${NAME_PRETTY} $NAME_VER" || DOT_COLOR="\e[1;92mONLINE \e[1;93m(non-systemd): ${NAME_PRETTY} $NAME_VER";;
 				*) DOT_COLOR="\e[1;93m???: ${NAME_PRETTY} $NAME_VER";;
 			esac
+			# create alternate screen buffer + hide cursor (only if first time)
 			clear
-			echo -e "$STATS"
 			printf "\n\e[${WATCH_LINES};0H\e[1;97m[${DOT_COLOR}\e[1;97m] [\e[1;95m%s\e[1;97m%s\e[1;94m%s\e[1;97m%s\e[0;97m%s\e[1;97m%s\e[0;97m%s\e[1;97m%s\e[0m " \
 				"$STAT_DATE" "] [" "$STAT_UPTIME" "] [${WATCH_REFRESH_RATE} sec] [" "$STAT_AMOUNT" "]"
 			# exit on any input unless [left] or [right] escape codes
@@ -165,10 +163,8 @@ watch_Status() {
 	else
 		COL="\e[1;92m"
 	fi
-	# create alternate screen buffer
-	printf "\e[?1049h"
-	# hide user input
-	stty -echo
+	# create alternate screen buffer + hide cursor + hide user input
+	[[ $FIRST_WATCH = true ]] && stty -echo && printf "\e[?1049h"
 	trap 'stty echo; printf "\e[?1049l\e[1;97m%s\e[1;95m%s\e[1;97m%s\n" "[Exiting: " "monero-bash watch" "]"; exit 0' EXIT
 	while :; do
 		[[ $XMRIG_VER ]] && sudo -v
@@ -184,8 +180,8 @@ watch_Status() {
 		if [[ $VAR_1 = $'\e' ]]; then
 			read -r -s -N 2 -t 1 VAR_2
 			case "$VAR_2" in
-				'[C') watch_Next ;;
-				'[D') watch_Prev ;;
+					'[C') watch_Next ;;
+					'[D') watch_Prev ;;
 				*) exit 0;;
 			esac
 		elif [[ $VAR_1 || $VAR_1 = $'\x0a' ]]; then